added kobayashi yields
Added Kobayashi+ 2006 SN yields, and the tools needed to read those files. These yields are useful as they have the final progenitor mass, allowing calculations of the mass lost to winds.

I cleaned up some of the read-in for Nomoto, as I added mass cut read in for those models as well.